Release/0.2

+++ src/main/java/com/spring/ApiSystem/domain/admin/AdminService.java
@@ -1,19 +1,25 @@
 package com.spring.ApiSystem.domain.admin;
 
 import com.spring.ApiSystem.domain.admin.dto.request.ReqAdicionarRoleDTO;
+import com.spring.ApiSystem.domain.admin.dto.request.ReqCadastroAdminDTO;
 import com.spring.ApiSystem.domain.admin.dto.request.ReqCadastroPersonalDTO;
 import com.spring.ApiSystem.domain.admin.dto.response.ResCadastrarPersonalDTO;
 import com.spring.ApiSystem.domain.admin.dto.response.ResRoleNeedDataDTO;
 import com.spring.ApiSystem.domain.admin.dto.response.ResUsuarioWithRolesResponseDTO;
 import com.spring.ApiSystem.domain.admin.exception.AdminNaoExisteException;
+import com.spring.ApiSystem.domain.aluno.Aluno;
 import com.spring.ApiSystem.domain.aluno.AlunoService;
 import com.spring.ApiSystem.domain.aluno.mapper.CpfMapper;
+import com.spring.ApiSystem.domain.personal.Personal;
 import com.spring.ApiSystem.domain.personal.PersonalService;
 import com.spring.ApiSystem.domain.telefone.Telefone;
 import com.spring.ApiSystem.domain.usuario.Usuario;
 import com.spring.ApiSystem.domain.usuario.UsuarioService;
 import com.spring.ApiSystem.domain.usuario.enums.Role;
 import com.spring.ApiSystem.domain.usuario.security.JpaUserDetailsService;
+import com.spring.ApiSystem.shared.infrastructure.email.dto.Email;
+import com.spring.ApiSystem.shared.infrastructure.email.service.EmailService;
+import com.spring.ApiSystem.shared.security.PasswordGenerator;
 import jakarta.transaction.Transactional;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -35,14 +41,16 @@ public class AdminService {
     private final JpaUserDetailsService userDetailsService;
     private final AlunoService alunoService;
     private final CpfMapper cpfMapper;
+    private final EmailService emailService;
 
-    public AdminService(AdminRepository adminRepository, UsuarioService usuarioService, PersonalService personalService, JpaUserDetailsService userDetailsService, AlunoService alunoService, CpfMapper cpfMapper) {
+    public AdminService(AdminRepository adminRepository, UsuarioService usuarioService, PersonalService personalService, JpaUserDetailsService userDetailsService, AlunoService alunoService, CpfMapper cpfMapper, EmailService emailService) {
         this.adminRepository = adminRepository;
         this.usuarioService = usuarioService;
         this.personalService = personalService;
         this.userDetailsService = userDetailsService;
         this.alunoService = alunoService;
         this.cpfMapper = cpfMapper;
+        this.emailService = emailService;
     }
 
     @Transactional
@@ -56,7 +64,7 @@ public class AdminService {
             throw new IllegalArgumentException("O usuário já possui a role: " + role);
         }
 
-        log.info("O Usuario id {} está adicionando role {} ao usuário de id {}", admin.getId(), role, userId);
+        log.info("O Admin de id {} está adicionando a role {} ao usuário de id {}", admin.getId(), role, userId);
 
         usuarioService.addRoleToUsuario(usuario, role);
 
@@ -112,10 +120,11 @@ public class AdminService {
         }
 
         usuarioService.salvarUsuario(usuario);
+        log.info("O Admin de id {} inativou (soft delete) o usuário de id {}", admin.getId(), userId);
     }
 
     @Transactional
-    public Usuario retirarRole(Role role, Long userId){
+    public Usuario removeRole(Role role, Long userId){
         Admin admin = userDetailsService.getCurrentAdmin();
 
         Usuario usuario = usuarioService.buscarUsuarioPorId(userId);
@@ -128,16 +137,26 @@ public class AdminService {
             throw new IllegalArgumentException("Um administrador não pode remover sua propria role.");
         }
 
+        if(usuario.isAdmin()){
+            throw new IllegalArgumentException("Um admistrador não pode ter a role de admistrador removida, apenas a role DONO posso fazer isso.");
+        }
+
+        if(usuario.hasOnlyRole(role)){
+            throw new IllegalArgumentException("Não é permitido remove a ultima role do usuario. Caso deseje deletar-lo utilize o meio de deletar.");
+        }
+
         usuario = usuarioService.removeRoleFromUsuario(usuario, role);
+
+        log.info("O Admin de id {} removeu a role {} do usuário de id {}", admin.getId(), role, userId);
 
         return usuario;
     }
 
     @Transactional
-    public ResCadastrarPersonalDTO criarPersonal(ReqCadastroPersonalDTO personalDTO){
+    public ResCadastrarPersonalDTO createPersonalUser(ReqCadastroPersonalDTO personalDTO){
         Admin admin = userDetailsService.getCurrentAdmin();
         ResCadastrarPersonalDTO personalCreated = personalService.cadastrarPersonalDto(personalDTO);
-        log.info("O Usuario id {} criou o personal de id {}", admin.getId(), personalCreated.id());
+        log.info("O Admin de id {} criou o personal de id {}", admin.getId(), personalCreated.id());
         return personalCreated;
     }
 
@@ -201,7 +220,7 @@ public class AdminService {
     }
 
     @Transactional
-    public Page<ResUsuarioWithRolesResponseDTO> listarUsuariosComFiltros(String nome, String email, Role role, Pageable pageable) {
+    public Page<ResUsuarioWithRolesResponseDTO> listUsersWithFilter(String nome, String email, Role role, Pageable pageable) {
         userDetailsService.getCurrentAdmin();
         Page<Usuario> usuarios;
         if (role != null) {
@@ -214,7 +233,7 @@ public class AdminService {
             String anamnese = null;
             String cref = null;
             if (usuario.isAluno() && usuario.getAluno() != null) {
-                var aluno = usuario.getAluno();
+                Aluno aluno = usuario.getAluno();
                 if (aluno.getCpf() != null) {
                     cpf = aluno.getCpf().toString();
                 }
@@ -223,7 +242,7 @@ public class AdminService {
                 }
             }
             if (usuario.isPersonal() && usuario.getPersonal() != null) {
-                var personal = usuario.getPersonal();
+                Personal personal = usuario.getPersonal();
                 if (personal.getCref() != null) {
                     cref = personal.getCref();
                 }
@@ -248,11 +267,49 @@ public class AdminService {
     }
 
     @Transactional
-    public void createAdminUser(String email, String password) {
+    public void createAdminUser(ReqCadastroAdminDTO dto) {
+        Admin admin = userDetailsService.getCurrentAdmin();
+        usuarioService.validarEmailExistente(dto.email());
+
+        Usuario adminCriado = new Usuario();
+        adminCriado.setNome(dto.nome());
+        adminCriado.setSexo(dto.sexo());
+        adminCriado.setRoles(Set.of(Role.ADMIN));
+        adminCriado.setDataNascimento(dto.dataNascimento());
+        adminCriado.setEmail(dto.email());
+        adminCriado.setTelefones(List.of(new Telefone(null, dto.telefone().pais(), dto.telefone().ddd(), dto.telefone().numero(), adminCriado)));
+
+        String randomSenha = PasswordGenerator.generate(10);
+        usuarioService.aplicarSenhaCriptografada(adminCriado, randomSenha);
+
+        adminCriado = usuarioService.salvarUsuario(adminCriado);
+        adminCriado = createProfile(adminCriado).getUsuario();
+
+        String corpoEmail = String.format(
+                "Olá %s,<br>Seu perfil de Admin foi criado com sucesso!<br><br>Sua senha temporária é: <strong>%s</strong>",
+                adminCriado.getNome(),
+                randomSenha
+        );
+
+        emailService.enviarEmail(new Email(
+                adminCriado.getEmail(),
+                "Bem-vindo ao sistema de academia",
+                corpoEmail
+        ));
+
+        log.info("O Admin de id {} criou um novo admin de id {}", admin.getId(), adminCriado.getId());
+    }
+
+    @Transactional
+    public void createInitialAdminUser(String email, String password) {
+        if(usuarioService.emailExiste(email)){
+            log.info("Admin inicial já existe, pulando criação.");
+            return;
+        }
         Usuario admin = new Usuario();
         usuarioService.aplicarSenhaCriptografada(admin, password);
         admin.setEmail(email);
-        admin.setNome("Fábio");
+        admin.setNome("ADMIN");
         admin.setSexo("M");
         admin.setRoles(Set.of(Role.ADMIN, Role.PERSONAL, Role.DONO));
         admin.setDataNascimento(LocalDate.parse("1990-01-01"));
@@ -261,6 +318,7 @@ public class AdminService {
         admin = usuarioService.salvarUsuario(admin);
         admin = createProfile(admin).getUsuario();
         personalService.createProfile(admin, "CREF12345");
-        log.info("Usuario admin inicial criado.");
+        log.info("Usuario admin inicial criado na inicialização.");
     }
 }
+
